Adds handling of edge cases regarding confusion matrices
When the confusion matrix has no false positives (i.e. filled only on the diagonal), the posterior
for a class is simply 1 because the evidence equals the likelihood * prior.

When the confusion matrix has all-zero columns (i.e. the inner classifier fails to make any
prediction for a given class), the posterior for these classes equals their prior probability.
This is necessary to make sure that the posterior probabilities of all classes, given a
prediction, sum to 1.