subjective classification (#213)
* Adds initial computation of posterior class probabilities, given prior and confusion matrix

* Adds tests for single posterior computations

The other ones are to make sure all posterior class probabilities add up to 1 for a given class prediction.

* Parameterizes test, adds several edge cases

... and breaks the posterior computation - it doesn't like all-zero columns in the confusion matrix.

* Typo

* Adds test assuring fit() stores confusion matrix as dataframe

* Prevents pain from missing classes in prior

* Adds handling of edge cases regarding confusion matrices

When the confusion matrix has no false positives (i.e. filled only on the diagonal), the posterior
for a class is simply 1 because the evidence equals the likelihood * prior.

When the confusion matrix has all-zero columns (i.e. the inner classifier fails to make any
prediction for a given class), the posterior for these classes equals their prior probability.
This is necessary to make sure that the posterior probabilities of all classes, given a
prediction, sum to 1.

* Adds predict_proba() implementation

* Adds guard against invalid model configuration

* ok, ok, numerical edge cases

* Flips expected/actual in tests to better understand failed tests

* Corrects docstring

* Adds sklearn boilerplate + checks

* Adds check for inner estimator to be fitted

* Adds evidence parameter to meta estimator

* Adds weighing of posterior by 'probas' of inner estimator

* Adds documentation for expected result of different evidence types

* Adds helper to convert probas to discrete predictions

* Adds optimization for predictions

Instead of computing the posterior for every observation at predict_proba(), the possible
posterior combinations (n_classes ** 2) are computed during fit() and stored in self as
posterior_matrix_. During prediction, this matrix only needs to be multiplied with the
(discrete) predictions, preventing lotsa recomputations of the same posteriors.

* Moves param checking to fit()

* Removes unnecessary references to inner estimator's classes_ attribute

It already can be referenced from the meta estimator's property.

* Adds docstring

* Moves method docstring to correct location