replace the example in `When Writing a New Feature` from `RandomAdder` to `PandasTypeSelector`
The RandomAdder was a buit strange as it uses the `TrainOnlyTransformerMixin` and therefore had a `transform_train` method instead of the more common `transform`.

Also, I took the opportunity to add a few remarks to the code indicating some of the oddities you need to be aware of when writing proper `scikit-learn` estimators.